Move store-related functionality to store dir
