+++ src/store/load-data.test.js
@@ -1,8 +1,8 @@
 import loadData from './load-data';
-import formatData from '../../utils/format-data';
-import loremIpsum from '../../utils/data/lorem-ipsum.mock';
-import animals from '../../utils/data/animals.mock';
-import demo from '../../utils/data/demo.mock';
+import formatData from '../utils/format-data';
+import loremIpsum from '../utils/data/lorem-ipsum.mock';
+import animals from '../utils/data/animals.mock';
+import demo from '../utils/data/demo.mock';
 
 describe('loadData', () => {
   it('returns the correct dataset when passed a dataset string', () => {
